+++ src/main/java/com/mamiyaotaru/voxelmap/gui/GuiButtonRowListPlayers.java
@@ -23,7 +23,7 @@ import java.util.Optional;
 
 public class GuiButtonRowListPlayers extends EntryListWidget<GuiButtonRowListPlayers.Row> {
     private final ArrayList<PlayerListEntry> players;
-    private ArrayList<?> playersFiltered;
+    private ArrayList<PlayerListEntry> playersFiltered;
     final GuiSelectPlayer parentGui;
     final Row everyoneRow;
     static final Text ALL = Text.translatable("minimap.waypointshare.all");
@@ -47,7 +47,7 @@ public class GuiButtonRowListPlayers extends EntryListWidget<GuiButtonRowListPla
     }
 
     private Text getPlayerName(PlayerListEntry ScoreboardEntryIn) {
-        return ScoreboardEntryIn.getDisplayName() != null ? ScoreboardEntryIn.getDisplayName() : Text.literal(ScoreboardEntryIn.getProfile().getName());
+        return Text.literal(ScoreboardEntryIn.getProfile().getName());
     }
 
     @NotNull
@@ -90,22 +90,25 @@ public class GuiButtonRowListPlayers extends EntryListWidget<GuiButtonRowListPla
         this.addEntry(this.everyoneRow);
 
         for (int i = 0; i < this.playersFiltered.size(); i += 2) {
-            PlayerListEntry ScoreboardEntry1 = (PlayerListEntry) this.playersFiltered.get(i);
-            PlayerListEntry ScoreboardEntry2 = i < this.playersFiltered.size() - 1 ? (PlayerListEntry) this.playersFiltered.get(i + 1) : null;
+            PlayerListEntry ScoreboardEntry1 = this.playersFiltered.get(i);
+            PlayerListEntry ScoreboardEntry2 = i < this.playersFiltered.size() - 1 ? this.playersFiltered.get(i + 1) : null;
             ButtonWidget guibutton1 = this.createButtonFor(this.parentGui.getWidth() / 2 - 155, 0, ScoreboardEntry1).get();
-            ButtonWidget guibutton2 = this.createButtonFor(this.parentGui.getWidth() / 2 - 155 + 160, 0, ScoreboardEntry2).get();
-            this.addEntry(new Row(guibutton1, i, guibutton2, i + 1));
+            if (ScoreboardEntry2 != null) {
+                ButtonWidget guibutton2 = this.createButtonFor(this.parentGui.getWidth() / 2 - 155 + 160, 0, ScoreboardEntry2).get();
+                this.addEntry(new Row(guibutton1, i, guibutton2, i + 1));
+            } else {
+                this.addEntry(new Row(guibutton1, i));
+            }
         }
-
     }
 
     public void buttonClicked(int id) {
         if (id == -1) {
             this.parentGui.allClicked = true;
-            ConfirmScreen confirmScreen = new ConfirmScreen(this.parentGui, this.TITLE, this.EXPLANATION, this.AFFIRM, this.DENY);
+            ConfirmScreen confirmScreen = new ConfirmScreen(this.parentGui, TITLE, EXPLANATION, AFFIRM, DENY);
             this.client.setScreen(confirmScreen);
         } else {
-            PlayerListEntry ScoreboardEntry = (PlayerListEntry) this.playersFiltered.get(id);
+            PlayerListEntry ScoreboardEntry = this.playersFiltered.get(id);
             String name = this.getPlayerName(ScoreboardEntry).getString();
             this.parentGui.sendMessageToPlayer(name);
         }
@@ -158,7 +161,7 @@ public class GuiButtonRowListPlayers extends EntryListWidget<GuiButtonRowListPla
         }
 
         private void drawIconForButton(MatrixStack matrixStack, ButtonWidget button, int id) {
-            PlayerListEntry networkPlayerInfo = (PlayerListEntry) GuiButtonRowListPlayers.this.playersFiltered.get(id);
+            PlayerListEntry networkPlayerInfo = GuiButtonRowListPlayers.this.playersFiltered.get(id);
             GameProfile gameProfile = networkPlayerInfo.getProfile();
             PlayerEntity entityPlayer = VoxelConstants.getPlayer().world.getPlayerByUuid(gameProfile.getId());
             RenderSystem.setShader(GameRenderer::getPositionTexProgram);